ejercicios_basicos/working_libreries/requirements.txt
numpy
pandas
whatismyip
varname
textblob
requests_cache
opencv-python
bcrypt
icecream
pycrypto